+++ longrunning/Cargo.toml
@@ -14,3 +14,4 @@ log = "0.4"
 tokio = { version = "1.7" }
 google-cloud-googleapis = { version = "0.2.0", path = "../googleapis" }
 google-cloud-gax = { version = "0.2.0", path = "../gax" }
+google-cloud-grpc= { version = "0.1.0", path = "../grpc" }
