+++ longrunning/src/autogen/mod.rs
@@ -1,0 +1,1 @@
+pub mod conn_pool;
