+++ spanner/src/admin/mod.rs
@@ -1,3 +1,7 @@
-pub mod conn_pool;
 pub mod database;
 pub mod instance;
+
+const SCOPES: [&str; 2] = [
+    "https://www.googleapis.com/auth/cloud-platform",
+    "https://www.googleapis.com/auth/spanner.admin",
+];
