+++ spanner/src/apiv1/conn_pool.rs
@@ -2,7 +2,7 @@ use google_cloud_auth::token_source::TokenSource;
 use google_cloud_auth::{create_token_source, Config};
 use google_cloud_googleapis::spanner::v1::spanner_client::SpannerClient;
 use google_cloud_grpc::conn::{
-    ConnectionManager as InternalConnectionManager, Error as InternalConnectionError,
+    ConnectionManager as GRPCConnectionManager, Error,
 };
 use std::sync::Arc;
 
@@ -15,32 +15,14 @@ const SCOPES: [&str; 2] = [
     "https://www.googleapis.com/auth/spanner.data",
 ];
 
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error(transparent)]
-    AuthInitialize(#[from] google_cloud_auth::error::Error),
-
-    #[error(transparent)]
-    InternalConnection(#[from] InternalConnectionError),
-}
-
 pub struct ConnectionManager {
-    inner: InternalConnectionManager,
-    token_source: Option<Arc<dyn TokenSource>>,
+    inner: GRPCConnectionManager,
 }
 
 impl ConnectionManager {
     pub async fn new(pool_size: usize, emulator_host: Option<String>) -> Result<Self, Error> {
         Ok(ConnectionManager {
-            inner: InternalConnectionManager::new(pool_size, SPANNER, AUDIENCE, emulator_host)
-                .await?,
-            token_source: Some(Arc::from(
-                create_token_source(Config {
-                    audience: Some(AUDIENCE),
-                    scopes: Some(&SCOPES),
-                })
-                .await?,
-            )),
+            inner: GRPCConnectionManager::new(pool_size, SPANNER, AUDIENCE, Some(&SCOPES), emulator_host).await?,
         })
     }
 
@@ -49,13 +31,7 @@ impl ConnectionManager {
     }
 
     pub fn conn(&self) -> Client {
-        //clone() reuses http/s connection
-        Client::new(
-            SpannerClient::new(self.inner.conn()),
-            match self.token_source.as_ref() {
-                Some(s) => Some(Arc::clone(s)),
-                None => None,
-            },
-        )
+        let (conn, ts) = self.inner.conn();
+        Client::new(SpannerClient::new(conn), ts)
     }
 }
